// src/libext/anchorme.d.ts
/*
 * Copyright (C) 2017 ZeXtras S.r.l.
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU General Public License
 * as published by the Free Software Foundation, version 2 of
 * the License.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License.
 * If not, see <http://www.gnu.org/licenses/>.
 */

// tslint:disable:interface-over-type-literal

export declare const anchorme: (str: string, opt?: IAnchormeOptions) => string | string[];

interface IAnchormeOptions {
  list?: boolean;
  attributes?: Array<AnchormeAttributeOption|AnchormeAttributeFcn>;
  emails?: boolean;
  urls?: boolean;
  ips?: boolean;
  files?: boolean;
}

export interface IAnchormeUrl {
  reason: "file"|"url"|"ip"|"email";
  protocol: "mailto:"|string;
  encoded: string;
  raw: string;
}

export type AnchormeAttributeFcn = (urlObj: IAnchormeUrl) => AnchormeAttributeOption;

export type AnchormeAttributeOption =
    { name: "target", value: "_blank" }
  | { name: "onclick", value: string }
  ;